Added GetNextIdentity() to IdentityQueueManager

+++ BrainThud.Web/Model/MasterConfiguration.cs
@@ -5,6 +5,6 @@ namespace BrainThud.Web.Model
     public class MasterConfiguration : TableServiceEntity
     {
         public int LastUsedUserId { get; set; }
-        public long CurrentMaxIdentity { get; set; }
+        public int CurrentMaxIdentity { get; set; }
     }
 }
